refactor: create specific file for event-bus-proxy
